chore(project-structure): reorganize folder structure
