+++ client/src/components/sprints/CheckboxList.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Grid } from 'semantic-ui-react';
 import Checkbox from './Checkbox';
 import GhostbusterButton from './GhostbusterButton';
-import { PaddedGrid } from './Styles/TeamStyles';
+import { PaddedGrid } from '../Styles/TeamStyles';
 
 const CheckboxList = props => {
   const { repos, handleCheckboxChange, storeCheckedRepos } = props;
