+++ client/src/components/sprints/MilestoneCommitMessages.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { List } from 'semantic-ui-react';
-import { StyledHeader, StyledList } from './Styles/TeamStyles';
-import sprints from '../../../server/config/sprints';
+import { StyledHeader, StyledList } from '../Styles/TeamStyles';
+import sprints from '../../../../server/config/sprints';
 
 const MilestoneCommitMessages = props => {
   const { sprint } = props;
